requirements.txt
python-telegram-bot>=20.7
openai>=1.12.0
notion-client>=2.2.1
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.3


